resolve finalizer naming issues
